chore: update axoproject and use changelog api
fixes #262
